Add sim_utils for commonly used remove_all_objects and make_sensor_specs functions in python tests and tutorials
