+++ examples/settings.py
@@ -4,6 +4,7 @@
 
 import habitat_sim
 import habitat_sim.agent
+from habitat_sim.utils import sim_utils as sut
 
 default_sim_settings = {
     # settings shared by example.py and benchmark.py
@@ -85,46 +86,11 @@ def make_cfg(settings):
         },
     }
 
-    # create sensor specifications
-    sensor_specs = []
-    for sensor_uuid, sensor_params in sensors.items():
-        if settings[sensor_uuid]:
-            # Check if type VisualSensorSpec
-            if (
-                sensor_params["sensor_type"] == habitat_sim.SensorType.COLOR
-                or sensor_params["sensor_type"] == habitat_sim.SensorType.DEPTH
-                or sensor_params["sensor_type"] == habitat_sim.SensorType.SEMANTIC
-            ):
-                # Check if type CameraSensorSpec
-                if (
-                    sensor_params["sensor_subtype"] == habitat_sim.SensorSubType.PINHOLE
-                    or sensor_params["sensor_subtype"]
-                    == habitat_sim.SensorSubType.ORTHOGRAPHIC
-                ):
-                    sensor_spec = habitat_sim.CameraSensorSpec()
-                    sensor_spec.uuid = sensor_uuid
-                    sensor_spec.sensor_type = sensor_params["sensor_type"]
-                    sensor_spec.sensor_subtype = sensor_params["sensor_subtype"]
-                    sensor_spec.resolution = sensor_params["resolution"]
-                    sensor_spec.position = sensor_params["position"]
-                    sensor_spec.gpu2gpu_transfer = False
-                    if not settings["silent"]:
-                        print("==== Initialized Sensor Spec: =====")
-                        print("Sensor uuid: ", sensor_spec.uuid)
-                        print("Sensor type: ", sensor_spec.sensor_type)
-                        print("Sensor position: ", sensor_spec.position)
-                        print("===================================")
-                    sensor_specs.append(sensor_spec)
-                else:
-                    # TODO: Add more checks for other types of sensors
-                    continue
-            else:
-                # TODO: Add more checks for NonVisualSensorSpec
-                continue
-
     # create agent specifications
     agent_cfg = habitat_sim.agent.AgentConfiguration()
-    agent_cfg.sensor_specifications = sensor_specs
+    agent_cfg.sensor_specifications = sut.make_sensor_specs_from_settings(
+        sensors, settings
+    )
     agent_cfg.action_space = {
         "move_forward": habitat_sim.agent.ActionSpec(
             "move_forward", habitat_sim.agent.ActuationSpec(amount=0.25)
