+++ examples/tutorials/nb_python/ECCV_2020_Interactivity.py
@@ -61,6 +61,7 @@ from PIL import Image
 
 import habitat_sim
 from habitat_sim.utils import common as ut
+from habitat_sim.utils import sim_utils as sut
 from habitat_sim.utils import viz_utils as vut
 
 try:
@@ -121,18 +122,21 @@ def make_cfg(settings):
             "resolution": [settings["height"], settings["width"]],
             "position": [0.0, settings["sensor_height"], 0.0],
             "orientation": [settings["sensor_pitch"], 0.0, 0.0],
+            "sensor_subtype": habitat_sim.SensorSubType.PINHOLE,
         },
         "depth_sensor_1st_person": {
             "sensor_type": habitat_sim.SensorType.DEPTH,
             "resolution": [settings["height"], settings["width"]],
             "position": [0.0, settings["sensor_height"], 0.0],
             "orientation": [settings["sensor_pitch"], 0.0, 0.0],
+            "sensor_subtype": habitat_sim.SensorSubType.PINHOLE,
         },
         "semantic_sensor_1st_person": {
             "sensor_type": habitat_sim.SensorType.SEMANTIC,
             "resolution": [settings["height"], settings["width"]],
             "position": [0.0, settings["sensor_height"], 0.0],
             "orientation": [settings["sensor_pitch"], 0.0, 0.0],
+            "sensor_subtype": habitat_sim.SensorSubType.PINHOLE,
         },
         # configure the 3rd person cam specifically:
         "color_sensor_3rd_person": {
@@ -140,24 +144,15 @@ def make_cfg(settings):
             "resolution": [settings["height"], settings["width"]],
             "position": [0.0, settings["sensor_height"] + 0.2, 0.2],
             "orientation": np.array([-math.pi / 4, 0, 0]),
+            "sensor_subtype": habitat_sim.SensorSubType.PINHOLE,
         },
     }
 
-    sensor_specs = []
-    for sensor_uuid, sensor_params in sensors.items():
-        if settings[sensor_uuid]:
-            sensor_spec = habitat_sim.CameraSensorSpec()
-            sensor_spec.uuid = sensor_uuid
-            sensor_spec.sensor_type = sensor_params["sensor_type"]
-            sensor_spec.resolution = sensor_params["resolution"]
-            sensor_spec.position = sensor_params["position"]
-            sensor_spec.orientation = sensor_params["orientation"]
-
-            sensor_specs.append(sensor_spec)
-
     # Here you can specify the amount of displacement in a forward action and the turn angle
     agent_cfg = habitat_sim.agent.AgentConfiguration()
-    agent_cfg.sensor_specifications = sensor_specs
+    agent_cfg.sensor_specifications = sut.make_sensor_specs_from_settings(
+        sensors, settings
+    )
 
     return habitat_sim.Configuration(sim_cfg, [agent_cfg])
 
@@ -202,14 +197,8 @@ def make_simulator_from_settings(sim_settings):
 # @title Define Simulation Utility Functions { display-mode: "form" }
 # @markdown (double click to show code)
 
-# @markdown - remove_all_objects
 # @markdown - simulate
 # @markdown - sample_object_state
-
-
-def remove_all_objects(sim):
-    for obj_id in sim.get_existing_object_ids():
-        sim.remove_object(obj_id)
 
 
 def simulate(sim, dt=1.0, get_frames=True):
@@ -431,7 +420,7 @@ def make_clear_all_objects_button():
         return
 
     def on_clear_click(b):
-        remove_all_objects(sim)
+        sut.remove_all_objects(sim)
 
     clear_objs_button = set_button_launcher("Clear all objects")
     clear_objs_button.on_click(on_clear_click)
@@ -603,7 +592,7 @@ build_widget_ui(obj_attr_mgr, prim_attr_mgr)
 # %%
 # @title Scripted vs. Dynamic Motion { display-mode: "form" }
 # @markdown A quick script to generate video data for AI classification of dynamically dropping vs. kinematically moving objects.
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 # @markdown Set the scene as dynamic or kinematic:
 scenario_is_kinematic = True  # @param {type:"boolean"}
 
@@ -634,14 +623,14 @@ if make_video:
         open_vid=show_video,
     )
 
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 
 # %%
 # @title Object Permanence { display-mode: "form" }
 # @markdown This example script demonstrates a possible object permanence task.
 # @markdown Two objects are dropped behind an occluder. One is removed while occluded.
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 # @markdown 1. Add the two dynamic objects.
 # add the selected objects
@@ -700,7 +689,7 @@ if make_video:
         output_path + example_type,
         open_vid=show_video,
     )
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 
 # %%
@@ -712,7 +701,7 @@ remove_all_objects(sim)
 
 introduce_surface = True  # @param{type:"boolean"}
 
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 # add a rolling object
 obj_attr_mgr = sim.get_object_template_manager()
@@ -747,7 +736,7 @@ if make_video:
         output_path + example_type,
         open_vid=show_video,
     )
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 
 # %%
@@ -760,7 +749,7 @@ remove_all_objects(sim)
 # @markdown Configure Parameters:
 
 obj_attr_mgr = sim.get_object_template_manager()
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 seed = 2  # @param{type:"integer"}
 random.seed(seed)
@@ -842,7 +831,7 @@ if make_video:
         output_path + example_type,
         open_vid=show_video,
     )
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 
 # %% [markdown]
 # ## Generating Scene Clutter on the NavMesh
@@ -897,7 +886,7 @@ obj_attr_mgr.register_template(sel_obj_template_cpy, "scaled_sel_obj")
 
 # add the selected object
 sim.navmesh_visualization = True
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 fails = 0
 for _obj in range(num_objects):
     obj_id_1 = sim.add_object_by_handle("scaled_sel_obj")
@@ -933,7 +922,7 @@ if make_video:
         output_path + example_type,
         open_vid=show_video,
     )
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
 sim.navmesh_visualization = False
 
 # %% [markdown]
@@ -1363,4 +1352,4 @@ if make_video:
 
 # remove locobot while leaving the agent node for later use
 sim.remove_object(locobot_id, delete_object_node=False)
-remove_all_objects(sim)
+sut.remove_all_objects(sim)
